Fixed some functions and tweaked quiet mode
Tweaked how quiet mode works by removing the prin() function and instead
implement the --log-level flag which dictates how much output is sent to
the terminal. Using the new funciton print_log() and the log_level
variable we can dictate how much output is sent. Thus a log_level of 0
is the same as parsing the --quiet flag and a log_level of 3 is the same
as parsing --verbose.

Thus this adds the ability to add a minimal output option, which is
log_level of 1. Having a log_level of 4 allows bash to parse "set -x"
which is used for debugging purposes.

Improved messages for parallelisation, now will show number of threads
when running.

Other changes include changing arithmetic conditionals, version bump,
and further breaking down of the main function into different functions.